refactoring error validation code
